+++ RELEASE_NOTES.md
@@ -2,8 +2,9 @@
 
 ## in develop
 
-* Increases number of retries when downloading manifests
 * Uses manifest files for large subjob inputs
+* Uses dxda to bulk-download manifest files
+* Increases number of retries when downloading single manifest files
 * Fixes error when parsing a field name with multiple `stage-*` prefixes
 * Allows file-to-string coercion for WDL inputs